% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{training_blocks}
\alias{training_blocks}
\title{Create Training Blocks}
\usage{
training_blocks(training_dates, actual_TL, diff_dates)
}
\arguments{
\item{training_dates}{training dates}

\item{actual_TL}{position of the actual training load}

\item{diff_dates}{difference in days}
}
\description{
Create Training Blocks
}
