% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_load.R
\docType{data}
\name{training_load}
\alias{training_load}
\title{Training load dataframe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84 rows and 5 columns.
}
\usage{
data("training_load", package = "ACWR")
}
\description{
A dataframe with the training load of 3 subjects.
}
\section{Variables}{

\describe{
  \item{ID}{ID of the subjects}
  \item{Week}{training weeks}
  \item{Day}{training days}
  \item{TL}{training load (arbitrary units)}
  \item{Training_Date}{training dates}
}
}

\keyword{datasets}
